requirements.txt
flask
PyPDF2
langgraph

langchain-google-vertexai
python-dotenv
google-cloud-aiplatform
